finish POST a JSON object into DB in azure

+++ CourseProvider.Infrastructure/Models/CourseUpdateRequest.cs
@@ -10,7 +10,7 @@ public class CourseUpdateRequest
     public string[]? Categories { get; set; }
     public string? Title { get; set; }
     public string? Ingress { get; set; }
-    public int StarRating { get; set; }
+    public decimal StarRating { get; set; }
     public string? Reviews { get; set; }
     public string? LikesInProcent { get; set; }
     public string? Likes { get; set; }
